refactored everything so it is easier to deal with for the TAs :)
