+++ .gitmodules
@@ -1,3 +1,1 @@
-[submodule "SoOpMissionSimulator"]
-	path = SoOpMissionSimulator
-	url = https://github.itap.purdue.edu/RadioNavigationLab/SoOpMissionSimulator.git
+
